Move test to standalone class
